requirements.txt
fastapi
uvicorn
python-multipart
python-telegram-bot
supabase
python-dotenv
